+++ lib/vote.php
@@ -26,7 +26,8 @@ class TK_GVote
      */
     public function __construct($post_id = null)
     {
-        global $wpdb;
+    	global $wpdb;
+		
         $this->wpdb = $wpdb;
 		$this->wpdb->enable_nulls = true;
         
@@ -49,18 +50,50 @@ class TK_GVote
     }
 
     /**
-     * @todo этот метод лишний в классе
      * @param integer $post_id
      * @return bool
      */
     public static function exists($post_id)
     {
-        global $wpdb;
+    	global $wpdb;
+		
         if (isset($post_id)) {
             
             $res = $wpdb->get_var(
                 $wpdb->prepare("SELECT count(id) FROM {$wpdb->prefix}tkgp_votes  WHERE post_id = %d;", array($post_id))
             );
+            if ($res) {
+                return true;
+            }
+        }
+        return false;
+    }
+	
+	/**
+     * Retun TRUE if Project is approved, else return FALSE.
+	 * 
+     * @return bool
+     */
+    public function approved()
+    {
+        if (!empty($this->vote_id) && !empty($this->project_id)) {
+            $query = '';
+			
+			if($this->variantExists()) { //если пользовательское голосование
+				$query = $this->wpdb->prepare("SELECT 1 FROM DUAL WHERE 0"); //SQL заглушка
+			} else {
+				$query = $this->wpdb->prepare("SELECT 1 FROM DUAL 
+                				WHERE 
+                				(SELECT COUNT(id) FROM {$this->wpdb->prefix}tkgp_usersvotes WHERE vote_id = %d AND variant_id = -1 GROUP BY variant_id) 
+                				> 
+                				(SELECT target_votes FROM {$this->wpdb->prefix}tkgp_votes WHERE id = %d AND post_id = %d)"
+                				, array($this->vote_id,
+										$this->vote_id,
+										$this->project_id)
+								);
+			}
+			
+            $res = $this->wpdb->get_var($query);
             if ($res) {
                 return true;
             }
@@ -94,32 +127,44 @@ class TK_GVote
                 'label' => _x('Vote for approval', 'Project Settings', 'tkgp'),
                 'desc' => _x('Number of votes in which the project is considered approved.', 'Project Settings',
                     'tkgp'),
-                'id' => 'tkgp_target_votes',
+                'id' => 'tkgp_vote_target_votes',
                 'type' => 'number',
+                'value' => 100,
                 'options' => array(
-                    'min' => '1',
-                    'value' => 100,
-                    'step' => 1
+                    'min' => 1,
+                    'step' => 1,
+                    'required' => 'true'
                 )
             ),
             array(
                 'label' => _x('Start date', 'Project Settings', 'tkgp'),
                 'desc' => _x('The date of commencement of voting.', 'Project Settings', 'tkgp'),
-                'id' => 'tkgp_start_date',
+                'id' => 'tkgp_vote_start_date',
                 'type' => 'date',
                 'options' => array('required')
             ),
             array(
                 'label' => _x('End date', 'Project Settings', 'tkgp'),
                 'desc' => _x('The data of the end of voting.', 'Project Settings', 'tkgp'),
-                'id' => 'tkgp_end_date',
+                'id' => 'tkgp_vote_end_date',
                 'type' => 'date'
+            ),
+            array(
+                'label' => _x('Allow re-vote', 'Project Settings', 'tkgp'),
+                'desc' => _x('Users can reset their vote and vote again, or not to vote :)', 'Project Settings', 'tkgp'),
+                'id' => 'tkgp_vote_allow_revote',
+                'type' => 'checkbox',
+                'exclude' => 1,
+                'value' => 1
             ),
             array(
                 'label' => _x('Reset voting results', 'Project Settings', 'tkgp'),
                 'desc' => _x('Reset voting results.', 'Project Settings', 'tkgp'),
-                'id' => 'tkgp_reset_vote',
-                'type' => 'checkbox'
+                'id' => 'tkgp_vote_reset',
+                'type' => 'checkbox',
+                'exclude' => 1,
+                'value' => 1
+                
             )
         );
     }
@@ -237,10 +282,11 @@ class TK_GVote
                     case 'start_date':
                     case 'end_date':
                     case 'target_votes':
-                        if ($columns == '*') {
+					case 'allow_revote':
+                        if ($columns === '*') {
                             $columns = $cur;
                         } else {
-                            $columns = $columns . ', ' . $columns;
+                            $columns .= ', ' . $cur;
                         }
                         break;
 
@@ -248,8 +294,6 @@ class TK_GVote
                         continue;
                 }
             }
-
-
 
             return $this->wpdb->get_row($this->wpdb->prepare("SELECT {$columns} FROM {$this->wpdb->prefix}tkgp_votes WHERE id = %d;",
                 $this->vote_id), ARRAY_A);
@@ -271,6 +315,7 @@ class TK_GVote
 				switch ($key) {
 					case 'enabled':
 					case 'target_votes':
+					case 'allow_revote':
 						$cur_format = '%d';
 					case 'start_date':
 					case 'end_date':
@@ -367,7 +412,10 @@ class TK_GVote
     public function deleteUserVote($user_id)
     {
         if (isset($this->vote_id) && isset($user_id)) {
-            
+        	$vote_settings = $this->getVoteSettings(array('allow_revote'));
+            if(!$vote_settings['allow_revote']) {
+            	return false;
+            }
 
             $res = $this->wpdb->query($this->wpdb->prepare("
 					DELETE FROM {$this->wpdb->prefix}tkgp_usersvotes 
@@ -377,12 +425,28 @@ class TK_GVote
                 $this->vote_id, $user_id)
             );
 
-            if (intval($res) == 0) {
+            if (intval($res) > 0) {
                 return true;
             }
         }
         return false;
     }
+
+	/**
+	 * Reset all votes
+	 */
+	public function resetVote() {
+		if (isset($this->vote_id)) {
+            
+
+            $res = $this->wpdb->query($this->wpdb->prepare("
+					DELETE FROM {$this->wpdb->prefix}tkgp_usersvotes 
+					WHERE vote_id = %d
+					",
+                $this->vote_id)
+            );
+        }
+	}
 
     /**
      * @param mixed[] $arg
@@ -441,9 +505,10 @@ class TK_GVote
         if (isset($this->vote_id)) {
 
 
-            $query = $this->wpdb->prepare("SELECT variant_id, variant 
+            $query = $this->wpdb->prepare("SELECT variant_id, variant, approval_flag 
 										FROM {$this->wpdb->prefix}tkgp_votevariant
 										WHERE vote_id = %d
+										ORDER BY approval_flag DESC
 										",
                 $this->vote_id);
 
@@ -454,45 +519,168 @@ class TK_GVote
     }
 
     /**
-     * @return string
-     */
-    public function getResultVoteHtml()
+     * When $show_vote_button = TRUE and $user_can_vote = FALSE then button "Reset my vote" displayed.
+	 * 
+	 * @param bool $show_vote_button[optional] Default is FALSE. Set TRUE when you need show vote buttons.
+	 * @param bool $short[optional] Defailt is FALSE. Set TRUE when you need show only Vote status.
+	 * @param null|bool $user_can_vote[optional] Default is TRUE. Set this param TRUE when current user can vote and.
+	 * @return string
+	 */
+    public function getResultVoteHtml($show_vote_button = false, $short = false, $user_can_vote = null)
     {
         $form = '';
 
         if (isset($this->vote_id)) {
-            $target_votes = $this->getVoteSettings(array('target_votes'));
-            $target_votes = floatval($target_votes['target_votes']);
-            $votes = $this->getVoteState();
+        	$settings = $this->getVoteSettings(array('target_votes','enabled','allow_revote','start_date','end_date'));
+            
             $form .= '<div id="tkgp_vote_result">
-            	<div><b>'. _x('Voting status', 'tk_project', 'tkgp') .'</b></div>';
+            	<div class="tkgp_title"><b>'. _x('Voting status', 'Project Vote', 'tkgp') .'</b></div>';
 
-            if ($this->variantExists()) {
-
-            } else {
-                $approval = 100.0 * floatval(isset($votes[0]) ? $votes[0]['cnt'] : 0) / $target_votes;
-                $reproval = 100.0 * floatval(isset($votes[1]) ? $votes[1]['cnt'] : 0) / $target_votes;
-
-                $form .= '<table>
-                	<tr><th>'. _x('Progress of the approval', 'tk_project', 'tkgp') .'</th>
-                	<td>
-                		<div id="tkgp_approval_status">';
-                $form .= '<div id="tkgp_approval" style="float: left; width: ' . $approval . '%; height: 100%; background: #FF0000;"></div>';
-                	/*Код на будущее, когда будет реализация голосов против Проекта*/
-					/*<div id="tkgp_reproval" style="float: left; width: ' . $reproval . '%; height: 100%; background: #EEE;"></div>';*/
-                $form .= '</td>
-                </tr>';
-				
-				$form .= '<tr><th>'. _x('Supported', 'tk_project', 'tkgp') . '</th><td>' . intval(isset($votes[0]) ? $votes[0]['cnt'] : 0) .'</td></tr>';
-				$form .= '<tr><th>'. _x('Not Supported', 'tk_project', 'tkgp') . '</th><td>' . intval(isset($votes[0]) ? $votes[1]['cnt'] : 0) .'</td></tr>';
-            	$form .= '</table>';
+            if ($this->variantExists()) { // формируем HTML для голосования с пользовательскими вариантами
+            	$form .= $this->getVariantVoteHtml($settings, $show_vote_button, $short, $user_can_vote);	
+            } else { // формируем HTML для стандартного голосования-одобрения проекта
+            	$form .= $this->getStandartVoteHtml($settings, $show_vote_button, $short, $user_can_vote);
 			}
 
             $form .= '</div>';
         }
         return $form;
     }
+	
+	/**
+	 * @param array $settings
+	 * @param bool $show_vote_button
+	 * @param bool $short
+	 * @param bool $user_can_vote
+	 * @return string
+	 */
+	protected function getStandartVoteHtml($settings, $show_vote_button, $short, $user_can_vote) {
+		$target_votes = floatval($settings['target_votes']);
+		$allow_revote = (bool)$settings['allow_revote'];
+		$votes = $this->getVoteState();
+		$approval_votes = 0;
+		$reproval_votes = 0;
 
+        if(isset($votes[0]) && intval($votes[0]['id']) === -1) {
+        	$approval_votes = intval($votes[0]['cnt']);
+			$reproval_votes = empty($votes[1]['cnt']) ? 0 : intval($votes[1]['cnt']);
+        } else {
+        	$reproval_votes = intval($votes[0]['cnt']);
+			$approval_votes = empty($votes[1]['cnt']) ? 0 : intval($votes[1]['cnt']);
+        }
+		
+		$approval_color = $approval_votes < $target_votes ? '#FF0000' : '#00FF00';
+		
+		$total_factor = $approval_votes > $target_votes ? intval($approval_votes / $target_votes) + 1 : 1; //Если голосов одобрения больше или равно требуемому количеству	
+		$total_votes = $approval_votes + $reproval_votes; //Всего проголосовало человек
+        $approval = 100.0 * floatval($approval_votes) / ($target_votes * $total_factor); //Процент одобрения от требуемых голосов для одобрения
+        $reproval = 100.0 * floatval($reproval_votes) / ($target_votes * $total_factor); //Процент не одобрения от требуемых голосов для одобрения
+               
+        $total_approval = $total_votes ? 100.0 * floatval($approval_votes) / $total_votes : 0; //Процент одобривших от общего количества проголосовавших
+		$total_reproval = $total_votes ? 100.0 * floatval($reproval_votes) / $total_votes : 0; //Процент не одобривших от общего количества проголосовавших
+		$approval_percent = str_replace(',0', '', number_format($total_approval, 1, ',', '')); //Процент одобривших от общего количества проголосовавших
+		$reproval_percent = str_replace(',0', '', number_format($total_reproval, 1, ',', '')); //Процент не одобривших от общего количества проголосовавших
+
+        $form .= '<table>
+        	<tr><th>'. _x('Progress of the approval', 'Project Vote', 'tkgp') .'</th>
+               		<td>
+                		<div id="tkgp_approval_status">';
+        $form .= '<div id="tkgp_approval" style="float: left; width: ' . $approval . '%; height: 100%; background: ' . $approval_color . ';"></div>';
+        /*Код на будущее, когда будет реализация голосов против Проекта*/
+		/*<div id="tkgp_reproval" style="float: left; width: ' . $reproval . '%; height: 100%; background: #EEE;"></div>';*/
+		$form .= '</div><div id="tkgp_approval_desc">' . ($approval_votes < $target_votes ? $approval_votes . ' <b>/</b> ' . $target_votes 
+															: _x('APPROVED!', 'Project Vote', 'tkgp'));
+        $form .= '</div></td>
+        </tr>';
+		if(!$short) {
+			$form .= '<tr><th>'. _x('Supported', 'Project Vote', 'tkgp') . '</th><td>' . intval($approval_votes) .' (' . $approval_percent . '%)</td></tr>';
+			$form .= '<tr><th>'. _x('Not Supported', 'Project Vote', 'tkgp') . '</th><td>' . intval($reproval_votes) .' (' . $reproval_percent . '%)</td></tr>';
+
+			if($show_vote_button && ($user_can_vote || $allow_revote)) {
+				$form .= '<tr><td colspan="2">' . ($user_can_vote ? $this->getVoteButtonHtml() : ($allow_revote ? $this->getVoteResetButtonHtml() : '')) . '</td></tr>';
+			}
+		}
+		$form .= '</table>';
+		
+		return $form;
+	}
+	
+	/**
+	 * @param array $settings
+	 * @param bool $show_vote_button
+	 * @param bool $short
+	 * @param bool $user_can_vote
+	 * @return string
+	 */
+	protected function getVariantVoteHtml($settings, $show_vote_button, $short, $user_can_vote) {
+		$form .= '';
+		
+		if($show_vote_button) {
+			$form .= $this->getVoteButtonHtml(true);
+		}
+		
+		return $form;
+	}
+	
+	/**
+	 * @param bool $variant_exists[optional]
+	 * @return string
+	 */
+	protected function getVoteButtonHtml($variant_exists = false) {
+		$html_code = '<div id="tkgp_vote_buttons">
+							<input type="hidden" name="tkgp_vote_nonce" value="'. wp_create_nonce('tkgp_user_vote') .'" />
+							<input type="hidden" name="tkgp_vote_id" value="'. $this->vote_id .'">
+							<input type="hidden" name="tkgp_post_id" value="'. $this->project_id .'">';
+		
+		if($variant_exists) {
+			$vars = $this->getVoteVariants();
+			$html_code .= '<ul class="tkgp_variants">';
+			
+			foreach ($vars as $var) {
+				$html_code .= '<li><input type="radio" name="user_vote" value="'. $var['variant_id'] .'" require/>'. $var['variant'] .'</li>';
+			}
+			
+			$html_code .= '</ul>
+			<div class="tkgp_button"><a>'. _x('Vote', 'Project Vote', 'tkgp') .'</a></div>';
+			
+		} else {
+			$html_code .= '<div>
+							<div class="tkgp_button">
+							<a>'. _x('I support','Project Vote', 'tkgp') .'</a>
+							<input type="hidden" name="user_vote" value="-1"/>
+						  	</div>
+						  </div>
+						  <div>
+						  	<div class="tkgp_button">
+						  		<a>'. _x('I do\'t support','Project Vote', 'tkgp') .'</a>
+						  		<input type="hidden" name="user_vote" value="-2"/>
+						  	</div>
+						  </div>';
+		}
+		
+		$html_code .= '</div>';
+		
+		return $html_code;
+	}
+	
+	/**
+	 * @return string
+	 */
+	protected function getVoteResetButtonHtml() {
+		$html_code = '<div>
+							<input type="hidden" name="tkgp_vote_nonce" value="'. wp_create_nonce('tkgp_reset_user_vote') .'" />
+							<input type="hidden" name="tkgp_vote_id" value="'. $this->vote_id .'">
+							<input type="hidden" name="tkgp_post_id" value="'. $this->project_id .'">
+							<div>
+								<div class="tkgp_button_reset">
+								<a>'. _x('Reset my vote','Project Vote', 'tkgp') .'</a>
+						  		</div>
+						  </div>
+					  </div>';
+							
+		return $html_code;
+	}
+	
     /**
      * @param mixed[] $arg
      * @return bool
@@ -532,6 +720,20 @@ class TK_GVote
         return false;
     }
 
+	/**
+	 * @return null|int
+	 */
+	public function getPostId() {
+		return $this->project_id;
+	}
+	
+	/**
+	 * @return null|int
+	 */
+	public function getVoteId() {
+		return $this->vote_id;
+	}
+	
     /**
      * @param mixed|mixed[] $val
      * @param mixed|mixed[] $default
